backend/requirements.txt
# Django Core
Django==5.1.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
django-cors-headers==4.9.0
django-filter==24.3
drf-spectacular==0.27.2

# Database
psycopg2-binary==2.9.11
django-redis==5.4.0
redis==5.0.1

# Celery
celery==5.3.4
django-celery-results==2.5.1
django-celery-beat==2.7.0
sqlalchemy==2.0.23

# Production
gunicorn==21.2.0
whitenoise==6.6.0
python-json-logger==2.0.7
django-environ==0.12.0

# HTTP Client
httpx==0.26.0
requests==2.32.3

# ASGI
asgiref==3.10.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2

# Testing and Quality
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
ruff==0.1.9
mypy==1.8.0
django-stubs==4.2.7
safety==2.3.5
bandit==1.7.6
locust==2.20.0
radon==6.0.1

# Monitoring
sentry-sdk==1.39.2
